Started modularizing login page, prevented multiple login attempts from bugging the code
