+++ client/src/index.js
@@ -51,17 +51,13 @@ root.render(
           <Route element={<Dashboard />} path="/dashboard" exact></Route>
           <Route element={<ListView />} path="/list-view" exact></Route>
           <Route element={<FullView />} path="/full-view/:_id" exact>
-            {" "}
           </Route>
           <Route element={<RecordPage />} path="/add-artefact" exact>
-            {" "}
           </Route>
           <Route element={<EditPage />} path="/edit-artefact/:_id" exact>
-            {" "}
           </Route>
 
           <Route element={<Skewer />} path="skewer" exact>
-            {" "}
           </Route>
         </Route>
       </Routes>
